% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ShinyTester.R
\name{ShinyDummyCheck}
\alias{ShinyDummyCheck}
\title{ShinyDummyCheck}
\usage{
ShinyDummyCheck(ui = "ui.R", server = "server.R", directory = getwd())
}
\arguments{
\item{ui}{a character vector size 1 containing the name of the UI files. defaults to "ui.R"}

\item{server}{a character vector size 1 containing the names of the SERVER file. defaults to "server.R"}

\item{directory}{the directory or website containing the files for the Shiny App. Defaults to current working directory}
}
\value{
Returns a dataframe with the matchings b/w ui and server files. Also spawns them in VIEW mode.
}
\description{
This function takes Shiny files themselves as inputs and tries to combine the different
assets presented in the ui and server files to see whether they match up.
}
\details{
For now, it only works where the server and ui files are seperate (ie, it doesn't work for `app.R` yet)
}
\examples{
ShinyDummyCheck(directory = "https://raw.githubusercontent.com/mexindian/ShinyServer/master/LineSelector")

## Or, to test with your own app, go to your shiny app, make that your working directory, and then type `ShinyDummyCheck()`
}

