% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ShinyTester.R
\name{ShinyHierarchy}
\alias{ShinyHierarchy}
\title{ShinyHierarchy}
\usage{
ShinyHierarchy(ui = "ui.R", server = "server.R", directory = getwd())
}
\arguments{
\item{ui}{a character vector size 1 containing the name of the UI files. defaults to "ui.R"}

\item{server}{a character vector size 1 containing the names of the SERVER file. defaults to "server.R"}

\item{directory}{the directory or website containing the files for the Shiny App}
}
\value{
It returns a very very nice network chart
}
\description{
Create a hierarchical network chart that  shows the _ad hoc_ structure of your shiny Server.
}
\examples{
ShinyHierarchy(directory = "https://raw.githubusercontent.com/mexindian/ShinyServer/master/LineSelector")

## Or, to test with your own app, go to your shiny app, make that your working directory, and then type `ShinyHierarchy()`
}

